% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_records.R
\name{eu_records}
\alias{eu_records}
\title{Get Europeana records}
\usage{
eu_records(ids, profile = NULL, hierarchy = NULL,
  key = getOption("eu_key"), ...)
}
\arguments{
\item{ids}{(character) Record ids}

\item{profile}{(character) Profile parameter controls the format and richness
of the response. One of full, similar, params.}

\item{hierarchy}{(character) One of 'self', 'parent', 'children', 'preceding-siblings',
'following-siblings', or 'ancestor-self-siblings'}

\item{key}{(character) API key for Europeana}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get Europeana records
}
\examples{
\dontrun{
eu_records('91916/C389BE4E26EC03D11E27C2F5D9CA209FD18D51EE')
eu_records('2020903/KMSB_37_106_7')
ids <- c('2020903/KMSB_37_106_7', '92099/BibliographicResource_2000081662431')
eu_records(ids)

# profile param
id <- '91916/C389BE4E26EC03D11E27C2F5D9CA209FD18D51EE'
eu_records(id, profile = "full")
eu_records(id, profile = "similar")
eu_records(id, profile = "params")

# should work even with ids that don't exist
eu_records(c('2020903/KMSB_37_106_7', 'notanid'))

# hierarchy
eu_records(ids = '9200300/BibliographicResource_3000116310788', hierarchy = "parent")
ids <- c('9200300/BibliographicResource_3000116310788',
   '9200300/BibliographicResource_3000116311348')
eu_records(ids, hierarchy = "parent")
eu_records(ids, hierarchy = "children")
}
}

