% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_suggestions.R
\name{eu_suggestions}
\alias{eu_suggestions}
\title{Get Europeana records}
\usage{
eu_suggestions(query, limit = 10, start = NULL, key = getOption("eu_key"),
  ...)
}
\arguments{
\item{query}{(character) Query string}

\item{key}{(character) API key for Europeana}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
DOESN"T WORK, THE SUGGESTIONS API SEEMS TO BE NOT WORKING
}
\examples{
\dontrun{
eu_suggestions(query = "war")
}
}

