% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{reuropeana-package}
\alias{reuropeana}
\alias{reuropeana-package}
\title{R Client for Europeana.}
\description{
R Client for Europeana
}
\details{
\tabular{ll}{
 Package: \tab reuropeana\cr
 Type: \tab Package\cr
 Version: \tab 0.0.1.99\cr
 Date: \tab 2014-10-28\cr
 License: \tab MIT\cr
}
}
\author{
Scott Chamberlain
}

